pyproject.toml
[project]
name = "secretgate"
version = "0.1.0"
description = "Lean security proxy for AI coding tools — scan and redact secrets before they reach LLM APIs"
requires-python = ">=3.11"
license = "Apache-2.0"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "httpx>=0.28",
    "pyyaml>=6.0",
    "click>=8.1",
    "structlog>=24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-httpx>=0.35",
    "ruff>=0.8",
]

[project.scripts]
secretgate = "secretgate.cli:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
